slist_r atomic version

+++ cpp11/courseSchedule.h
@@ -97,11 +97,11 @@ template <typename T>
 class cstring_iterator : public iterator<forward_iterator_tag, T>  // T is the underlying value type
 {
 	const char **str;
-	int count;
-	int increment;
+	size_t count;
+	size_t increment;
 	const char **end;
 public:
-	cstring_iterator(int c, const char *argv[], int inc) :str(argv), count(c), increment(inc), end(argv + c) {	}
+	cstring_iterator(size_t c, const char *argv[], size_t inc) :str(argv), count(c), increment(inc), end(argv + c) {	}
 	cstring_iterator() = default;// used as end()
 	bool operator!=(const cstring_iterator&) {  // clever trick, check if it is end, from left hand only
 		return str != end;
